[Model] Rename and add docs
