+++ gammagl/models/__init__.py
@@ -19,9 +19,10 @@ from .gprgnn import GPRGNNModel
 from .pna import PNAModel
 from .mixhop import MIXHOPModel
 from .hardgat import HardGATModel
-from .generator import Generator
-from .discriminator import Discriminator
 from .graphgan import GraphGAN
+from .graphgan_generator import Generator
+from .graphgan_discriminator import Discriminator
+from .pna import PNAModel
 
 __all__ = [
     'GCNModel',
@@ -46,9 +47,10 @@ __all__ = [
     'PNAModel',
     'MIXHOPModel',
     'HardGATModel',
+    'GraphGAN',
     'Generator',
     'Discriminator',
-    'GraphGAN'
+    'PNAModel',
 ]
 
 classes = __all__
