+++ gammagl/models/graphgan_discriminator.py
