+++ gammagl/models/graphgan_generator.py
